Refactor into Logic class
